+++ redux/longestStreakSlice.ts
@@ -1,9 +1,9 @@
 import { createSlice, PayloadAction, Slice } from "@reduxjs/toolkit";
 import { RootState } from "./rootReducer";
 import { BlockState } from "../components/math/longestStreak/MultiplicationBlock";
+import { GameBlockState, initializeGameState, } from "../pages/studentPortal/labs/multiplication/game";
 import { getRandomItemFromArray, getRndInteger } from "../pages/api/random";
-import { GameBlockState, GameLevelsMetaData, gameLevelsMetaData, initializeGameState } from "../pages/api/longestStreak";
-import { min } from "lodash";
+import { shuffle } from "lodash";
 
 export interface LongestStreakState {
   stage: STAGE;
@@ -13,6 +13,7 @@ export interface LongestStreakState {
   isPlayerSelecting: boolean;
   currentlySelectedBlock?: number;
   playerName: string;
+  level: GameLevel;
 }
 
 export enum STAGE {
@@ -23,19 +24,19 @@ export enum STAGE {
 
 
 export enum GameLevel {
-  BEGINNER = "beginner",
-  BEGINNER_ADVANCED = "beginnerAdvanced",
-  INTERMEDIATE = "intermediate",
-  INTERMEDIATE_ADVANCED = "intermediateAdvanced",
-  EXPERT = "expert",
+  BEGINNER,
+  BEGINNER_ADVANCED,
+  INTERMEDIATE,
+  INTERMEDIATE_ADVANCED,
+  EXPERT,
 }
 
-//function that takes in GameLevel as input
 
 const initialState: LongestStreakState = {
   stage: STAGE.SET_RULES,
   reset: false,
-  blocks: initializeGameState(GameLevel.BEGINNER),
+  level: GameLevel.BEGINNER,
+  blocks: initializeGameState(GameLevel.INTERMEDIATE),
   handlePlayerSelect: 0,
   isPlayerSelecting: false,
   currentlySelectedBlock: null,
@@ -45,8 +46,8 @@ const initialState: LongestStreakState = {
 const resetInitialState: LongestStreakState = {
   stage: STAGE.PLAY_GAME,
   reset: false,
-  blocks: initializeGameState(GameLevel.BEGINNER),
-
+  level: GameLevel.BEGINNER,
+  blocks: initializeGameState(GameLevel.INTERMEDIATE),
   handlePlayerSelect: 0,
   isPlayerSelecting: false,
   currentlySelectedBlock: null,
@@ -72,9 +73,8 @@ export const longestStreakSlice: Slice = createSlice({
       state.blocks = selectedBlock;
     },
 
-    initializeGame: (state: LongestStreakState, action: PayloadAction<GameLevel>) => {
-      const currentLevel = action.payload
-      state.blocks = initializeGameState(currentLevel);
+    initializeGame: (state: LongestStreakState, action: PayloadAction) => {
+      state.blocks = initializeGameState(state.level);
     },
 
 
@@ -101,6 +101,11 @@ export const longestStreakSlice: Slice = createSlice({
       }
     },
     
+    setLevel: (state: LongestStreakState, action: PayloadAction<GameLevel>) => {
+      const levelOfGame = action.payload as GameLevel;
+      state.level = levelOfGame;
+    },
+
     handlePlayerSelect: (state, action: PayloadAction<number>) => {
       if (action.type === "longestStreak/handlePlayerSelect") {
         const index = action.payload;
@@ -198,6 +203,7 @@ export const {
   initializeGame,
   currentlySelectedBlock,
   setPlayerName,
+  setLevel
 } = longestStreakSlice.actions;
 
 export const longestStreakSelector = (state: RootState) =>
